refactor dynamodb client
